For alternate sources use the full command name as the name
This makes it less ambiguous what command is being run. When running 'npm run dev' it is more obvious what was run if that's the name instead of just showing 'dev' in the list like before. This also makes it so for makefile sources we can show 'make' instead of '<default target>'. With more additional sources now it made the most sense to keep it unambiguous.