If the split window no longer has a terminal buffer in it, create a new one when toggled
If you change the contents of the window that has the terminal split in it, like if you do this:
1. open a command in a split
2. in the split open a file to edit
3. try and close the split using toggle_split

If you do that it was previously trying to close your former split window, even though it's not used by projectlaunch anymore. This meant you couldn't open a new split either! If you only have one window left you'll also just get an error saying you can't close the last remaining window.

This enhances the checks for if the split window is usable to not only check that the window is still valid, but also to check if it still has a terminal buffer in it. If it doesn't have a terminal buffer in it anymore it'll open a new split.