Pressing X on a command that has exited will now remove it from the list
If you run a command but don't want the command in your list of commands anymore, this lets you remove it.

This also fixes a bug where pressing a key in insert mode in a terminal that has exited, which closes the buffer, then trying to reopen a split/float for that command would throw an error.