Add an option to control if the split window is focused automatically when opening
If I open the split usually it's just to check something that's long running, I generally don't need to interact with it. The default behavior now will re-focus the window that was active before the split is toggled, but you can choose to pass a new option to `setup` which lets you re-enable the old behavior.