Added highlight groups, editable commands, expose api for custom commands
I rewrote what the main menu looks like when jobs are running. They now have distinct highlight groups, and I'm no longer trying to right justify the status text. With really long commands that overflowed the menu the status text looked kinda broken. Now there's no specific "running" status that shows, but if a command exits it shows the exit code and uses a highlight group based on comments so it'll likely have a muted tone.

Commands can now be edited with `e`. If a custom command is edited it updates it, if anything else is updated it creates a new custom command. Custom commands can also be added programmatically as a method to add a command is now exposed.